requirements.txt
Flask==2.2.2
python-telegram-bot==20.1
gallery-dl==1.22.0